+++ lib/reactive_forms.dart
@@ -13,6 +13,8 @@ export 'src/models/form_array.dart';
 export 'src/models/form_control.dart';
 export 'src/models/form_control_collection.dart';
 export 'src/models/form_group.dart';
+export 'src/utils/control_extensions.dart';
+export 'src/utils/control_utils.dart';
 export 'src/validators/validation_message.dart';
 export 'src/validators/validator.dart';
 export 'src/validators/validators.dart';