+++ lib/src/exceptions/form_control_not_found_exception.dart
@@ -5,13 +5,17 @@
 /// This exception is thrown when a [FormGroup] or a [FormArray]
 /// doesn't find the [FormControl] by name
 class FormControlNotFoundException implements Exception {
-  final String name;
+  final String controlName;
 
   /// Creates an instance of the exception
-  FormControlNotFoundException(this.name);
+  FormControlNotFoundException({this.controlName});
 
   @override
   String toString() {
-    return 'FormControlNotFoundException: control with name: \'$name\' not found.';
+    if (this.controlName == null) {
+      return 'FormControlNotFoundException: control not found.';
+    }
+
+    return 'FormControlNotFoundException: control with name: \'$controlName\' not found.';
   }
 }