+++ lib/src/widgets/reactive_status_listenable_builder.dart
@@ -22,7 +22,8 @@ class ReactiveStatusListenableBuilder extends StatelessWidget {
     @required this.builder,
     this.child,
   })  : assert(formControlName != null),
-        assert(builder != null);
+        assert(builder != null),
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {